refactor: reduce qsint schemes to 1
